requirements-tests.txt
# Testing tools
unittest2>=1.0.1
coverage<=3.7.1
flake8>=2.4.1
mock>=1.0.1
nose>=1.3.6
django-nose>=1.4
tox>=2.0.1
termcolor>=1.1.0
